+++ src/core/pipeline/items/logical/or.rs
@@ -21,10 +21,10 @@ impl OrModifier {
 impl Item for OrModifier {
     async fn call<'a>(&self, ctx: Ctx<'a>) -> Result<Ctx<'a>> {
         if ctx.value.is_null() {
-            Ok(ctx)
+            Ok(ctx.clone())
         } else {
             match &self.value {
-                Value::Pipeline(p) => Ok(ctx.with_value(p.process(ctx).await?)),
+                Value::Pipeline(p) => Ok(ctx.clone().with_value(p.process(ctx.clone()).await?)),
                 _ => Ok(ctx.with_value(self.value.clone())),
             }
         }