+++ src/core/pipeline/items/object/object_set_value.rs
@@ -24,6 +24,6 @@ impl Item for ObjectSetValueModifier {
         let key = self.key.resolve(ctx.clone()).await?;
         let value = self.value.resolve(ctx.clone()).await?;
         ctx.object.as_ref().unwrap().set_value(key.as_str().unwrap(), value).unwrap();
-        ctx
+        Ok(ctx)
     }
 }