+++ src/core/pipeline/items/object/object_value.rs
@@ -20,6 +20,6 @@ impl Item for ObjectValueModifier {
     async fn call<'a>(&self, ctx: Ctx<'a>) -> Result<Ctx<'a>> {
         let key = self.key.resolve(ctx.clone()).await?;
         let value = ctx.object.as_ref().unwrap().get_value(key.as_raw_enum_choice().unwrap()).unwrap();
-        ctx.with_value(value).with_path(path![key.as_raw_enum_choice().unwrap().to_string()])
+        Ok(ctx.with_value(value).with_path(path![key.as_raw_enum_choice().unwrap().to_string()]))
     }
 }