+++ src/core/pipeline/items/string/generation/random_digits.rs
@@ -22,6 +22,6 @@ impl RandomDigitsModifier {
 impl Item for RandomDigitsModifier {
     async fn call<'a>(&self, ctx: Ctx<'a>) -> Result<Ctx<'a>> {
         let len = self.len.resolve(ctx.clone()).await?;
-        ctx.with_value(Value::String(generate(len.as_usize().unwrap(), "1234567890")))
+        Ok(ctx.with_value(Value::String(generate(len.as_usize().unwrap(), "1234567890"))))
     }
 }